avifgainmaputil: Fix MSVC type conversion warnings
Fix MSVC C4244 warnings on type conversions, possible loss of data.

Do rounded division by using integer arithmetic.

+++ apps/avifgainmaputil/swapbase_command.cc
@@ -23,7 +23,7 @@ avifResult ChangeBase(const avifImage& image, int depth,
   swapped->yuvFormat = yuvFormat;
 
   const float headroom =
-      static_cast<double>(image.gainMap->metadata.alternateHdrHeadroomN) /
+      static_cast<float>(image.gainMap->metadata.alternateHdrHeadroomN) /
       image.gainMap->metadata.alternateHdrHeadroomD;
   const bool tone_mapping_to_sdr = (headroom == 0.0f);
